requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
bleach==6.0.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
cookiecutter==2.1.1
deprecation==2.1.0
Django==4.1.7
django-browser-reload==1.7.0
django-tailwind==3.5.0
docutils==0.19
dotty-dict==1.3.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
invoke==1.7.3
jaraco.classes==3.2.3
Jinja2==3.1.2
jinja2-time==0.2.0
keyring==23.13.1
MarkupSafe==2.1.2
more-itertools==9.1.0
multidict==6.0.4
openai==0.27.2
packaging==23.0
pip==23.0.1
pkginfo==1.9.6
postgrest==0.10.6
psycopg2-binary==2.9.5
pydantic==1.10.7
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==3.13.0
python-semantic-release==7.33.2
python-slugify==8.0.1
PyYAML==6.0
readme-renderer==37.3
realtime==1.0.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
semver==2.13.0
setuptools==60.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.3
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.2
supafunc==0.2.2
tenacity==8.2.2
text-unidecode==1.3
tomlkit==0.11.7
tqdm==4.65.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==1.26.15
webencodings==0.5.1
websockets==10.4
wheel==0.40.0
yarl==1.8.2
zipp==3.15.0
